server/requirements.txt
APScheduler==3.6.*
asgiref==3.3.*
astroid==2.5.*
certifi==2020.12.*
chardet==4.0.*
Django==3.*
django-cors-headers==3.7.*
django-filter==2.4.*
djangorestframework==3.12.*
docker==4.4.*
idna==2.10
isort==5.7.*
jsonfield2==4.0.*
lazy-object-proxy==1.5.*
mccabe==0.6.*
pylint==2.7.*
pylint-django==2.4.*
pylint-plugin-utils==0.6
python-telegram-bot==13.3
pytz==2021.*
psycopg2==2.8.*
requests==2.25.*
six==1.15.*
sqlparse==0.4.*
toml==0.10.*
tornado==6.1
tzlocal==2.1
urllib3==1.26.*
websocket-client==0.58.*
wrapt==1.12.*
